+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Alle wichtigen Änderungen dieses Projekts werden in dieser Datei dokumentiert.
 
 Das Format basiert auf [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+### Changed
+- Tools Page eingeblendet
+
 ## [1.1.6.4]
 
 ### Added